feat: event page
