+++ src/assets/scss/elements/_form.scss
@@ -98,8 +98,10 @@ textarea {
 		align-items: center;
 		justify-content: space-between;
 
-		&__end {
-			justify-self: end;
+		&-center {
+			display: flex;
+			align-items: center;
+			justify-content: center;
 		}
 	}
 }
