+++ src/components/RegistrationForm/Steps/Step1.jsx
@@ -1,124 +1,135 @@
 import React from "react";
+import SectionTitle from "../../../components/SectionTitles/SectionTitle";
 import PropTypes from "prop-types";
-import { Formik, Form, Field, ErrorMessage } from "formik";
-import * as yup from "yup";
 
-const validationSchema = yup.object().shape({
-	prefix: yup.string().required("Please enter your prefered prefix"),
-	fullName: yup.string().required("Please enter your full names"),
-	email: yup
-		.string()
-		.email("Please enter a valid email")
-		.required("Please enter your email"),
-	phoneNumber: yup.string().required("Please enter your phone number"),
-	position: yup.string().required("Please enter your position"),
-	personalBio: yup.string().required("Please enter your personal bio"),
-});
-
-const Step1 = ({ step, setStep, data, setData }) => {
+const Step1 = ({ step, setStep }) => {
 	console.log(step);
 	return (
-		<Formik
-			data-aos="fade-up"
-			initialValues={{
-				prefix: data.prefix,
-				fullName: data.fullName,
-				email: data.email,
-				phoneNumber: data.phoneNumber,
-				position: data.position,
-				personalBio: data.personalBio,
-			}}
-			validationSchema={validationSchema}
-			onSubmit={(values) => {
-				setData({ ...data, ...values });
-				setStep(2);
-			}}
-		>
-			<Form className="registration-form-div">
-				<h2
-					className={`title`}
-					dangerouslySetInnerHTML={{ __html: "Personal Information" }}
-				></h2>
-				<div className="registration-form-element">
-					<label htmlFor="prefix">Prefix</label>
-					<Field as="select" name="prefix">
-						<option value="">Choose a Prefix</option>
-						<option value="Mr">Mr</option>
-						<option value="Mrs">Mrs</option>
-						<option value="Ms">Ms</option>
-						<option value="Dr">Dr</option>
-						<option value="Prof">Prof</option>
-					</Field>
-					<ErrorMessage name="prefix">
-						{(error) => (
-							<p className="registration-form-error">{error}</p>
-						)}
-					</ErrorMessage>
-				</div>
-				<div className="registration-form-element">
-					<label htmlFor="fullName">Full Names</label>
-					<Field name="fullName" type="text" />
-					<ErrorMessage name="fullName">
-						{(error) => (
-							<p className="registration-form-error">{error}</p>
-						)}
-					</ErrorMessage>
-				</div>
-				<div className="registration-form-element">
-					<label htmlFor="email">Email Address</label>
-					<Field name="email" type="email" />
-					<ErrorMessage name="email">
-						{(error) => (
-							<p className="registration-form-error">{error}</p>
-						)}
-					</ErrorMessage>
-				</div>
-				<div className="registration-form-element">
-					<label htmlFor="phoneNumber">Position</label>
-					<Field name="position" type="text" />
-					<ErrorMessage name="position">
-						{(error) => (
-							<p className="registration-form-error">{error}</p>
-						)}
-					</ErrorMessage>
-				</div>
-				<div className="registration-form-element">
-					<label htmlFor="position">Phone number</label>
-					<Field name="phoneNumber" type="number" />
-					<ErrorMessage name="phoneNumber">
-						{(error) => (
-							<p className="registration-form-error">{error}</p>
-						)}
-					</ErrorMessage>
-				</div>
-				<div className="registration-form-element">
-					<label htmlFor="personalBio">Personal Bio</label>
-					<Field name="personalBio" as="textarea" type="text" />
-					<ErrorMessage name="personalBio">
-						{(error) => (
-							<p className="registration-form-error">{error}</p>
-						)}
-					</ErrorMessage>
-				</div>
-				<div className="registration-form-buttons">
-					<p></p>
-					<button
-						className="btn btn-primary btn-hover-secondary mt-6 registration-form-buttons__end"
-						type="submit"
+		<div className="video-section section section-padding">
+			<div className="container">
+				<SectionTitle title="Please got through the checklist below before you start filling the application form" />
+				<div className="row align-items-center">
+					<div
+						className="skill-with-video-content"
+						data-aos="fade-up"
 					>
-						Next
-					</button>
+						<div className="section-title-two mb-4">
+							<h3 className="title">
+								What to look for before form submission
+							</h3>
+						</div>
+
+						<ul className="agency-list">
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									Ensure that the email address and phone
+									numbers are correct before form submission
+								</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									Review your information for spelling/typing
+									errors
+								</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									Short Personal Bio (Not more than 200 words)
+								</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									After submitting the form you will be
+									directed to a page confirming a success
+									submission, if not please resend your
+									application
+								</div>
+							</li>
+						</ul>
+					</div>
+
+					<div
+						className="skill-with-video-content"
+						data-aos="fade-up"
+					>
+						<div className="section-title-two mb-4">
+							<h3 className="title">
+								Documents you will need to attach
+							</h3>
+						</div>
+
+						<ul className="agency-list">
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									An updated CV / Resume (PDF/DOC/DOCX)
+								</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									A signed national integrity Pledge
+								</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">Company Pofile (PDF)</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									Company Logo (JPEG/PDF)
+								</div>
+							</li>
+							<li className="item">
+								<div className="icon">
+									<i className="fas fa-check"></i>
+								</div>
+								<div className="text">
+									CEO&apos;s photo (JPEG not less than 500 X
+									500px)
+								</div>
+							</li>
+						</ul>
+					</div>
 				</div>
-			</Form>
-		</Formik>
+			</div>
+			<div className="registration-form-buttons-center mt-4">
+				<button
+					className="btn btn-primary btn-hover-secondary mt-6 registration-form-buttons__end"
+					type="button"
+					onClick={() => setStep(2)}
+				>
+					Start your application here
+				</button>
+			</div>
+		</div>
 	);
 };
 
 Step1.propTypes = {
 	step: PropTypes.number,
 	setStep: PropTypes.func,
-	data: PropTypes.object,
-	setData: PropTypes.func,
 };
 
 export default Step1;
