+++ src/components/RegistrationForm/index.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import Step1 from "./Steps/Step1";
 import Step2 from "./Steps/Step2";
 import Step3 from "./Steps/Step3";
-// import Step4 from "./Steps/Step4";
+import Step4 from "./Steps/Step4";
 
 const RegistrationForm = () => {
 	const [step, setStep] = useState(1);
@@ -36,14 +36,7 @@ const RegistrationForm = () => {
 	const steps = [
 		{
 			id: 1,
-			component: (
-				<Step1
-					data={data}
-					setData={setData}
-					step={step}
-					setStep={setStep}
-				/>
-			),
+			component: <Step1 step={step} setStep={setStep} />,
 		},
 		{
 			id: 2,
@@ -60,6 +53,17 @@ const RegistrationForm = () => {
 			id: 3,
 			component: (
 				<Step3
+					data={data}
+					setData={setData}
+					step={step}
+					setStep={setStep}
+				/>
+			),
+		},
+		{
+			id: 4,
+			component: (
+				<Step4
 					data={data}
 					setData={setData}
 					step={step}
