+++ src/container/ContactFromContainer/ContactFromContainer.js
@@ -10,8 +10,8 @@ const ContactFromContainer = () => {
                     <div className="offset-lg-2 col-lg-8">
                         <SectionTitle
                             headingOption="fz-32"
-                            title="Let’s talk about your project"
-                            subTitle="We have made it easy for clients to reach us and get their solutions weaved"
+                            title="Contact Us"
+                            subTitle="We have made it easy to reach us."
                         />
                         <ContactFrom />
                     </div>
