+++ src/data/contactInfo/contactInfo.json
@@ -15,6 +15,6 @@
 		"id": "3",
 		"icon": "images/svg/linea/linea-basic-mail-open-text.svg",
 		"title": "Email Us",
-		"info": "info@tanzaniteceoroundtable.co.tz <br/>  <p>&nbsp;</p>"
+		"info": "info@tanzaniteceoroundtable.co.tz <br/>  info@tanzaniteceoroundtable.com"
 	}
 ]
